+++ src/main/java/blackjack/common/GameResultType.java
@@ -4,6 +4,4 @@ public enum GameResultType {
     WIN,
     TIE,
     LOSE;
-
-
 }
