+++ src/main/java/blackjack/domain/CardHolder.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import java.util.List;
