+++ src/main/java/blackjack/domain/GameRule.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain;
 
 import java.util.List;
