+++ src/main/java/blackjack/manager/CardsGenerator.java
@@ -1,5 +1,8 @@
-package blackjack;
+package blackjack.manager;
 
+import blackjack.CardRank;
+import blackjack.CardSuit;
+import blackjack.domain.Card;
 import java.util.Arrays;
 import java.util.List;
