+++ src/test/java/blackjack/CardsGeneratorTest.java
@@ -2,6 +2,8 @@ package blackjack;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.Card;
+import blackjack.manager.CardsGenerator;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
