+++ src/test/java/blackjack/domain/CardTest.java
@@ -1,7 +1,9 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.CardRank;
+import blackjack.CardSuit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
