+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,11 +1,14 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import blackjack.CardRank;
+import blackjack.CardSuit;
 import blackjack.common.ErrorMessage;
+import blackjack.manager.CardsGenerator;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
