+++ src/test/java/blackjack/manager/BlackJackInitManagerTest.java
@@ -1,8 +1,10 @@
-package blackjack;
+package blackjack.manager;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.Hand;
+import blackjack.domain.Players;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
